Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -4,8 +4,8 @@
  * linear_search - searches for a value in an array of
  * integers using the Linear search algorithm
  *
- * @array: input array
- * @size: size of the array
+ * @array: pointer to first element of the array to search in 
+ * @size: number of elements in array
  * @value: value to search in
  * Return: Always EXIT_SUCCESS
  */
@@ -16,7 +16,7 @@ int linear_search(int *array, size_t size, int value)
 	if (array == NULL)
 		return (-1);
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (value == array[i])
